[FIX] purchase delivery views
