+++ purchase_delivery/model/__init__.py
@@ -5,4 +5,5 @@
 
 from . import delivery
 from . import purchase
+from . import purchase_line
 from . import stock
