+++ purchase_delivery/view/purchase_view.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <record id="view_order_withcarrier_form" model="ir.ui.view">
+        <field name="name">delivery.purchase.order_withcarrier.form.view</field>
+        <field name="model">purchase.order</field>
+        <field name="inherit_id" ref="purchase.purchase_order_form"/>
+        <field name="arch" type="xml">
+            <xpath expr="//sheet/notebook/page[2]/group/group[2]"  position="after">
+                <group name="delivery">
+                    <field name="carrier_id" context="{'purchase_order_id':active_id or False}"
+                            help="If you don't 'Add in Quote', the exact price will be computed when invoicing based on incoming shipment(s)."/>
+                    <button name="delivery_set" string="Add in Quote" type="object"
+                        class="oe_edit_only"
+                        attrs="{'invisible':[('carrier_id','=',False),('state','not in',('draft','sent'))]}"/>
+                </group>
+            </xpath>
+        </field>
+    </record>
+
+
+</odoo>
